Authorization implementation using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Question;
 use App\Http\Requests\AskQuestionRequest;
+use Illuminate\Support\Facades\Gate;
 
 class QuestionsController extends Controller
 {
@@ -65,6 +66,10 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403);
+        }
+
         return view('questions.edit', compact('question'));
     }
 
@@ -77,6 +82,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if (Gate::denies('update-question', $question)) {
+            abort(403);
+        }
+
         $question->update($request->only('title', 'text'));
 
         return redirect()->route('questions.index')->with('success', 'Your question has been updated');
@@ -91,6 +100,10 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
+        if (Gate::denies('delete-question', $question)) {
+            abort(403);
+        }
+
         $question->delete();
 
         return redirect()->route('questions.index')->with('success', 'Question has been successfully deleted');
